Got discord oauth working and users inserted into turso db
