+++ client/src/routes/signin.lazy.tsx
@@ -1,66 +1,13 @@
 import { createLazyFileRoute } from '@tanstack/react-router'
-import { useMutation, useQueryClient } from '@tanstack/react-query'
-import { useNavigate } from '@tanstack/react-router'
-import { useEffect } from 'react';
 
 export const Route = createLazyFileRoute('/signin')({
   component: SignIn,
 })
 
-async function initiateDiscordAuth() {
-  const response = await fetch('/api/auth/discord', { credentials: 'include'});
-  if(response.ok) {
-    const data = await response.json();
-    console.log(data)
-    window.location.href = data;
-  } else {
-    throw new Error('Failed to initiate Discord Authentication');
-  }
-}
-
-async function handleAuthCallback(code: string) {
-  const response = await fetch(`/api/auth/discord/callback?code=${code}`, {
-    method: 'GET',
-    credentials: 'include'
-  });
-  if(response.ok) {
-    return true;
-  } else {
-    throw new Error('Authentication failed')
-  }
-}
-
 export function SignIn() {
-  const queryClient = useQueryClient();
-  const navigate = useNavigate();
 
-  const authMutation = useMutation({
-    mutationFn: handleAuthCallback,
-    onSuccess: () => {
-      queryClient.invalidateQueries({ queryKey: ['user'] });
-      navigate({ to: '/' })
-    }
-  })
-
-  useEffect(() => {
-    const urlParams = new URLSearchParams(window.location.search)
-    const code = urlParams.get('code')
-    console.log(code)
-    if(code) {
-      authMutation.mutate(code)
-    }
-  }, [])
-
-  const handleSignIn = () => {
-    initiateDiscordAuth()
-  }
-
-  if (authMutation.isPending) {
-    return <div className="flex justify-center items-center h-screen">Authenticating...</div>;
-  }
-
-  if (authMutation.isError) {
-    return <div className="flex justify-center items-center h-screen text-red-500">Authentication failed. Please try again.</div>;
+  function handleSignIn(){
+    window.location.href = "https://discord.com/oauth2/authorize?client_id=1267324967679299664&response_type=code&redirect_uri=http%3A%2F%2Flocalhost%3A3000%2Fapi%2Fauth%2Fdiscord%2Fcallback&scope=identify";
   }
 
   return (
